\name{tr}
\alias{tr}
\title{Trace of a matrix.}
\description{Compute the trace of a square matrix.}
\usage{
  tr(X)
}
\arguments{
  \item{X}{Numeric: A square matrix.}
}
\details{
  This computes the sum of the diagonal elements of a square matrix.
}
\value{
  \item{numeric}{A real number.}
}
\author{
   Cedric E. Ginestet <cedric.ginestet@kcl.ac.uk>
}
